+++ include/allegro/fixed.h
@@ -25,7 +25,7 @@
    extern "C" {
 #endif
 
-typedef long fixed;
+typedef int32_t fixed;
 
 AL_VAR(AL_CONST fixed, fixtorad_r);
 AL_VAR(AL_CONST fixed, radtofix_r);
